Save functionality
Code that has two functions. One takes a character object and a filename and saves the info stored in the object to the file. The other takes a filename and saves the data from the file into the object and returns the object.
Used for saving game and loading a saved game.